Implement shortenurl process and update shortedurl columns
Add endpoints to generate shorted url and to retrieve it from shortcode,
the shortcode is generated in base of table primary key using HashidsNet
https://www.nuget.org/packages/Hashids.net/, the implementation of HashidsNet
is injected as a singleton service in ShortenerController.

Due of the primary key using in shortcode creation, the model ShortenedUrl
needed to be modify to avoid the id to be inncremented by the database, cause
we need to know the id to generate the shortcode, if the id is autoincremented
efcore throw exception if we provide a value to this field.