Fixing bugs in webclient shorten process
When shortening a link, if index was found in the application url,
this word was also included in the shortened link, because a header
called "referer" was used to take the domain name, which includes all
the current route at the time of the request (its mean https://localhost:443/Index),
it was changed and now it uses a concatenation of Request.Scheme and Request.Host.

Add javascript message when a badrequest is returned when consuming
api shorten endpoint.